Move auth template files into Organizations app
Similar to but subtly different from the changes in the previous commit,
the Accounts app cannot keep auth templates: Although at this time we only
have a single app Organizations that directly depends on Accounts,
Accounts is designed to be used by "many" apps that are of "main program"
nature (e.g. Admin, Lori, Ma-Service), each of which wishes to bring its
own templates. As such, Accounts acts as a third-party app that should
only keep what generalizes well (views, forms) to several dependent apps
that use it.