Distinguish between personal views and member views
Both are about views of user data that tend to be very similar to each
other – this is why is seemed self-suggesting to implement both via
common, shared views. For example, a common user overview was used both as
personal overview as well as member overview.

However, it turns out that that approach is both confusing and complicated:
Using „user“ as generic term to refer to both the `request.user` as well as
to other users (members) is very confusing.
Moreover, the views and templates for each tend to be a lot more different
than anticipated.
Clearly distinguishing between personal views and member views solves these
issues.